move search property types into types folder
